fixes, refactoring & documentation
